Moved all files to a new package call Activities. Add some testing for DB
