+++ app/src/main/java/dyly/bloomu/edu/gardenplannerapp/Activities/AddNewImageActivity.java
@@ -1,27 +1,26 @@
-package dyly.bloomu.edu.gardenplannerapp;
+package dyly.bloomu.edu.gardenplannerapp.Activities;
 
-import android.content.Intent;
+import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuItem;
-import android.view.View;
 
-/**
- * Created by EVANDESKTOP on 11/9/2015.
- */
-public class GardenNotesActivity extends AppCompatActivity {
+import dyly.bloomu.edu.gardenplannerapp.R;
+
+
+public class AddNewImageActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_garden_notes);
+        setContentView(R.layout.activity_add_new_image);
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_bed_notes_actvity, menu);
+        getMenuInflater().inflate(R.menu.menu_add_new_image, menu);
         return true;
     }
 
@@ -39,12 +38,4 @@ public class GardenNotesActivity extends AppCompatActivity {
 
         return super.onOptionsItemSelected(item);
     }
-
-    public void redirectToAddNewNoteActivity(View view)
-    {
-        Intent intent = new Intent(this, AddNewNoteActivity.class);
-        startActivity(intent);
-    }
-
-
 }
