+++ app/src/main/java/dyly/bloomu/edu/gardenplannerapp/Activities/BedLayoutActivity.java
@@ -1,4 +1,4 @@
-package dyly.bloomu.edu.gardenplannerapp;
+package dyly.bloomu.edu.gardenplannerapp.Activities;
 
 import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
@@ -6,19 +6,21 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuItem;
 
+import dyly.bloomu.edu.gardenplannerapp.R;
 
-public class AddNewImageActivity extends AppCompatActivity {
+
+public class BedLayoutActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_add_new_image);
+        setContentView(R.layout.activity_bed_layout);
     }
 
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         // Inflate the menu; this adds items to the action bar if it is present.
-        getMenuInflater().inflate(R.menu.menu_add_new_image, menu);
+        getMenuInflater().inflate(R.menu.menu_bed_layout, menu);
         return true;
     }
