+++ app/src/main/java/dyly/bloomu/edu/gardenplannerapp/Activities/BedOptionsActivity.java
@@ -1,4 +1,4 @@
-package dyly.bloomu.edu.gardenplannerapp;
+package dyly.bloomu.edu.gardenplannerapp.Activities;
 
 import android.content.Intent;
 import android.os.Bundle;
@@ -6,6 +6,8 @@ import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuItem;
 import android.view.View;
+
+import dyly.bloomu.edu.gardenplannerapp.R;
 
 
 public class BedOptionsActivity extends AppCompatActivity {
