+++ app/src/main/java/dyly/bloomu/edu/gardenplannerapp/Activities/CustomBedNotesHeader.java
@@ -1,8 +1,10 @@
-package dyly.bloomu.edu.gardenplannerapp;
+package dyly.bloomu.edu.gardenplannerapp.Activities;
 
 import android.content.Context;
 import android.view.LayoutInflater;
 import android.widget.LinearLayout;
+
+import dyly.bloomu.edu.gardenplannerapp.R;
 
 /**
  * Created by EVANDESKTOP on 11/5/2015.
