+++ app/src/main/java/dyly/bloomu/edu/gardenplannerapp/Activities/SupportActivity.java
@@ -1,10 +1,12 @@
-package dyly.bloomu.edu.gardenplannerapp;
+package dyly.bloomu.edu.gardenplannerapp.Activities;
 
 import android.support.v7.app.ActionBarActivity;
 import android.os.Bundle;
 import android.support.v7.app.AppCompatActivity;
 import android.view.Menu;
 import android.view.MenuItem;
+
+import dyly.bloomu.edu.gardenplannerapp.R;
 
 
 public class SupportActivity extends AppCompatActivity {
